==> old_notebooks/1_tratamiento_de_datos.ipynb <==
import os
import urllib3
import zipfile

import pandas as pd

# 1. Tratamiento de datos

En este notebook preparamos la información de usuarios, items y ratings que vamos a utilizar en el experimento. Para estos experimentos usaremos el dataset de MovieLens 100K y un dataset de IMDB con 5000 películas.

## 1.1. Preparación de entorno y Descarga de dataset

A continuación, preparamos el entorno para ejecutar los experimentos. Creamos las carpetas que vamos a necesitar en los siguientes notebooks.

# Creamos las carpetas que no existan

if not os.path.exists('data/experiment_data'):
    os.mkdir('data/experiment_data')

if not os.path.exists('data/matrices_data'):
    os.mkdir('data/matrices_data')
    
if not os.path.exists('data/similarity_data'):
    os.mkdir('data/similarity_data')
    
if not os.path.exists('data/results'):
    os.mkdir('data/results')

## Obtenemos los ratings del experimento

Ahora, cargamos todos los ratings y nos quedamos con aquellos de los usuarios con más actividad y de los items con más ratings.

all_ratings_DF = pd.read_csv('data/original_dataset/ml-latest-small/ratings.csv')
all_ratings_DF.head()

# Obtenemos los identificadores de los usuarios y las películas
users_id = all_ratings_DF.groupby(by='userId').count().sort_values('rating', ascending=False).index.values
movies_id = all_ratings_DF.groupby(by='userId').count().sort_values('rating', ascending=False).index.values

# Filtramos de los ratings aquellos usuarios e items que no están en las listas calculadas antes
final_ratings = all_ratings_DF[all_ratings_DF['userId'].isin(users_id)]
final_ratings = final_ratings[final_ratings['movieId'].isin(movies_id)]
final_ratings.head()

# Guardamos los ratings filtrados
final_ratings.to_csv('data/experiment_data/ratings.csv', index=False)

## Siguiente Notebook

Ya tenemos el dataset del experimento preparado. A continuación, hay que descargar toda la información posible de las películas. Para ello, ejecute el notebook `extraer_informacion_peliculas`
