==> old_notebooks/2_new_extraer_informacion_peliculas.ipynb <==
import pandas as pd

# 2. Extraer información de las películas

En este notebook vamos a extraer toda la información que podamos de las películas. Para ello, hemos descargado un dataset con toda la información de IMDB de 5000 películas. Este dataset se encuentra en https://www.kaggle.com/carolzhangdc/imdb-5000-movie-dataset.

## 2.1. Cargamos los datasets

Cargamos los datasets de las películas que existen en *MovieLens* y las películas que tenemos en el dataset de los metadatos de *IMBD*.

imdb_df = pd.read_csv('data/original_dataset/movie_metadata.csv')
imdb_df.head()

# Cargamos los links en IMDB de las películas
movielens_df = pd.read_csv('data/original_dataset/ml-latest-small/links.csv')
movielens_df.head()

# Filtramos las películas que hay en ratings
ratings_df = pd.read_csv('data/experiment_data/ratings.csv')
movielens_df = movielens_df[movielens_df['movieId'].isin(set(ratings_df['movieId']))]
movielens_df.head()

## 2.2. Obtener las películas que hay en ambos dataset

A contianuación, nos quedamos con las películas que se encuentran en ambos datasets. Para ello, vamos a obtener el identificador de la películas en *IMDB* y quedarnos con aquellas paelículas cuyo identificador de *IMDB* se encuentran en los dos datasets.

def get_id(row):
    '''
    Función para obtener ek identificador de IMDB
    a partir de su URL.
    '''
    return int(row[28:].split('/')[0])

# Creamos las columna 'imdbId' en el datasets de IMDB
imdb_df['movie_imdb_id'] = imdb_df.apply(lambda row: get_id(row['movie_imdb_link']), axis=1)

# Quitamos de los 2 datasets las películas que no están en alguno de los datasets
movielens_df = movielens_df[movielens_df['imdbId'].isin(imdb_df['movie_imdb_id'].values)]
imdb_df = imdb_df[imdb_df.movie_imdb_id.isin(movielens_df['imdbId'].values)]

# Eliminamos las filas duplicadas
imdb_df.drop_duplicates(subset=['movie_imdb_id'], inplace=True)

# Si el hay valor NaN lo convertimos a 0
imdb_df.fillna(0, inplace=True)
imdb_df.head()

## 2.3. Limpiamos datos

Eliminamos los datos del datasets de *IMDB* que no necesitamos en el experimento. Además, la información de la duración y el año los pasamos a valores enteros. Por último, al dataset de *IMDB* le añadimos los identificadores de *MovieLens* de las películas que serán con las que trabajaremos.

def get_actors(row):
    return str(row['actor_1_name']) + '|' + str(row['actor_3_name']) + '|' + str(row['actor_3_name'])

# Juntamos los actores
imdb_df['stars'] = imdb_df.apply(lambda row: get_actors(row), axis=1)

# Eliminamos columnas
columns_to_drop = ['color', 
                   'num_critic_for_reviews', 
                   'director_facebook_likes', 
                   'actor_3_facebook_likes', 
                   'actor_2_facebook_likes', 
                   'actor_1_facebook_likes',
                   'num_user_for_reviews',
                   'movie_facebook_likes',
                   'gross',
                   'num_voted_users',
                   'cast_total_facebook_likes',
                   'facenumber_in_poster',
                   'movie_imdb_link',
                   'budget',
                   'imdb_score',
                   'aspect_ratio',
                   'plot_keywords',
                   'actor_1_name',
                   'actor_2_name',
                   'actor_3_name',
                   'content_rating']

imdb_df.drop(columns=columns_to_drop, inplace=True)

# Convertimos los datos de las columnas años y duración
imdb_df['title_year'] = imdb_df.apply(lambda row: str(int(row['title_year'])), axis=1)
imdb_df['duration'] = imdb_df.apply(lambda row: str(int(row['duration'])), axis=1)

# Añadimos los identificadores de las películas
imdb_df.sort_values(by=['movie_imdb_id'], ascending=True, inplace=True)
movielens_df.sort_values(by=['imdbId'], ascending=True, inplace=True)
imdb_df['id'] = movielens_df['movieId'].values
imdb_df.set_index('id', inplace=True)
imdb_df.reset_index(inplace=True)

imdb_df.drop(columns=['movie_imdb_id'], inplace=True)
imdb_df.head()

## 2.4. Guardar datasets limpio

Guardamos el dataset en un fichero CSV para utilizar más adelante.

imdb_df.to_csv('data/experiment_data/movies.csv', index=False)

## 2.5. Filtrar peliculas en ratings CSV

Eliminamos las películas de ratings.csv que no están representadas en `movies_data_df`.

ratings_df = pd.read_csv('data/experiment_data/ratings.csv')
len(ratings_df)

ratings_df.head()

ratings_df = ratings_df[ratings_df['movieId'].isin(imdb_df.id.values)]
len(ratings_df)

ratings_df.to_csv('data/experiment_data/ratings.csv', index=False)

## Siguiente Notebook

Ya tenemos los datasets del experimento preparados. A continuación, hay que descargar toda la información posible de las películas. Para ello, ejecute el notebook `vectores_binarios_peliculas`
