==> old_notebooks/3_vectores_binarios_peliculas.ipynb <==
import pandas as pd
from tqdm import tqdm

# 3. Crear vectores binarios de películas

En este notebook, vamos a convertir las características de las películas en vectores binarios de características. Estos vectores se utilizarán más adelante para calcular el **FCA**.

def to_string(value):
    return str(value)

# Cargamos las 
movies_DF = pd.read_csv('data/experiment_data/movies.csv')
movies_DF['title_year'] = movies_DF['title_year'].apply(lambda val: to_string(val))
movies_DF.head()

def get_dummie(df, column, sep):
    new_df = df[column].str.get_dummies(sep=sep)
    result = pd.concat([df, new_df], axis=1)        
    #result.drop(columns=[column])
    return result

lista_columns = ['director_name', 'genres', 'stars', 'language', 'country', 'title_year']

for c in tqdm(range(len(lista_columns))):
    movies_DF = get_dummie(movies_DF, lista_columns[c], sep='|')

lista_columns_to_drop = ['director_name', 'genres', 'stars', 'language', 'country', 'title_year', 'movie_title', 'duration']
movies_DF.drop(columns=lista_columns_to_drop, axis=1, inplace=True)
movies_DF.head()

# Guardamos los vectores en un nuevo CSV
movies_DF.to_csv('data/experiment_data/movies_binary.csv', index=False)

### Siguiente Notebook

Una vez que hemos obtenido los vectores binarios, tenemos que entrenar el algoritmo de factorización de matrices y calcular la matriz $Q$ de cada usuario. El siguiente notebook que hay que ejecutar es `calcular_matrices_q.ipynb`.
