==> old_notebooks/5_calcular_matrices_similitud.ipynb <==
import pandas as pd
import numpy as np
import src.similarities as sim
from tqdm import tqdm

# 5. Calcular matrices de similitud

En este notebook calculamos las matrices de similitud que se utilizan en la evaluación del sistema. Cada una de las matrices se guarda en un fichero CSV.

## 5.1. Matriz de similitud de propiedades de películas

Creamos una matriz de similitud entre las películas sobre sus propiedades. Para la matriz de similitud se usa la función *Jaccard*.

functions = ['euclidean', 'cosine', 'manhattan']
for f in functions:
    if f == 'euclidean':
        matrix_binary_sim = sim.get_binary_similarity_matrix(sim.euclidean_sim)
        matrix_binary_sim.to_csv('data/similarity_data/sim_{}_binary.csv'.format(f), index = False)
    elif f == 'cosine':
        matrix_binary_sim = sim.get_binary_similarity_matrix(sim.cosine)
        matrix_binary_sim.to_csv('data/similarity_data/sim_{}_binary.csv'.format(f), index = False)
    elif f == 'manhattan':
        matrix_binary_sim = sim.get_binary_similarity_matrix(sim.manhattan_sim)
        matrix_binary_sim.to_csv('data/similarity_data/sim_{}_binary.csv'.format(f), index = False)

## 5.2. Matrices de similitud sobre Q

A continuación, calculamos las matrices de similitud de cada usuario aplicando su matriz Q. Se utilizarán las funciones de similitud `euclidean`, `cosine` y `manhattan`.

# Cargamos a los usuarios
ratings_DF = pd.read_csv('data/experiment_data/trainset.csv')
users = list(set(ratings_DF['userId']))
functions = ['euclidean', 'cosine', 'manhattan']

for u in tqdm(range(len(users))):
    for f in range(len(functions)):
        # Calculamos la matriz filtrada de similitud
        if functions[f] == 'euclidean':
            matrix = sim.get_similarity_matrix(users[u], sim.euclidean_sim)
            matrix.to_csv('data/similarity_data/sim_{}_user_{}.csv'.format(functions[f], users[u]), index=False)
        elif functions[f] == 'cosine':
            matrix = sim.get_similarity_matrix(users[u], sim.cosine_sim)
            matrix.to_csv('data/similarity_data/sim_{}_user_{}.csv'.format(functions[f], users[u]), index=False)
        elif functions[f] == 'manhattan':
            matrix = sim.get_similarity_matrix(users[u], sim.manhattan_sim)
            matrix.to_csv('data/similarity_data/sim_{}_user_{}.csv'.format(functions[f], users[u]), index=False)

## Siguiente Notebook

Ya tenemos las matrices de similitud preparadas. A continuación, hay que evaluar el sitema. Para ello, ejecute el notebook `evaluar_KNN`.
