==> old_notebooks/7_Calcular_Mejora.ipynb <==
import pandas as pd
import numpy as np

all_ratings_df = pd.read_csv('data/experiment_data/result_all_values.csv')
all_ratings_df.head()

for k in range(1, 11):
    all_ratings_df['error_binary_k_{}'.format(k)] = all_ratings_df.apply(lambda row: row['rating_binary_k_{}'.format(k)] - row['predicted'], axis=1)
    
for k in range(1, 11):
    all_ratings_df['error_q_k_{}'.format(k)] = all_ratings_df.apply(lambda row: row['rating_q_k_{}'.format(k)] - row['predicted'], axis=1)

k_values = [1, 2, 3, 5, 10]

for k in range(1, 11):
    all_ratings_df.drop(columns='rating_binary_k_{}'.format(k), inplace=True)
    all_ratings_df.drop(columns='rating_q_k_{}'.format(k), inplace=True)

all_ratings_df.drop(columns=['predicted', 'userId', 'movieId'], inplace=True)
all_ratings_df.head()

def sum_err(values, k):
    limit = num_err(values, k)
    sum_values = values[:limit]
    
    total = 0.0
    for v in sum_values:
        total += abs(v)
        
    return total

def num_err(values, k):
    valid_values = np.sum(~np.isnan(values))
    limit = min(k, valid_values)
            
    return limit

for k in k_values:
    all_ratings_df['sum_err_bin_k_{}'.format(k)] = all_ratings_df.apply(lambda row: sum_err(row.values[1:11], k), axis=1)
    all_ratings_df['num_err_bin_k_{}'.format(k)] = all_ratings_df.apply(lambda row: num_err(row.values[1:11], k), axis=1)
    
for k in k_values:
    all_ratings_df['sum_err_q_k_{}'.format(k)] = all_ratings_df.apply(lambda row: sum_err(row.values[11:21], k), axis=1)
    all_ratings_df['num_err_q_k_{}'.format(k)] = all_ratings_df.apply(lambda row: num_err(row.values[11:21], k), axis=1)

for k in range(1, 11):
    all_ratings_df.drop(columns='error_binary_k_{}'.format(k), inplace=True)
    all_ratings_df.drop(columns='error_q_k_{}'.format(k), inplace=True)

sum_values = pd.DataFrame(all_ratings_df.groupby(by='rating').sum())
sum_values.reset_index(inplace=True)
sum_values

for k in k_values:
    sum_values['mae_binary_k_{}'.format(k)] = sum_values.apply(lambda row: row['sum_err_bin_k_{}'.format(k)] / row['num_err_bin_k_{}'.format(k)], axis=1)
    sum_values['mae_q_k_{}'.format(k)] = sum_values.apply(lambda row: row['sum_err_q_k_{}'.format(k)] / row['num_err_q_k_{}'.format(k)], axis=1)

for k in k_values:
    sum_values.drop(columns='sum_err_bin_k_{}'.format(k), inplace=True)
    sum_values.drop(columns='sum_err_q_k_{}'.format(k), inplace=True)

for k in k_values[1:]:
    sum_values.drop(columns='num_err_bin_k_{}'.format(k), inplace=True)
    sum_values.drop(columns='num_err_q_k_{}'.format(k), inplace=True)

sum_values.drop(columns='num_err_q_k_1', inplace=True)
sum_values.rename(columns={'num_err_bin_k_1': 'N'}, inplace=True)
sum_values

for k in k_values:
    sum_values['improvement_k_{}'.format(k)] = sum_values.apply(lambda row: (row['mae_binary_k_{}'.format(k)] - row['mae_q_k_{}'.format(k)]) / row['mae_binary_k_{}'.format(k)] * 100, axis=1)

sum_values.to_csv('data/results/mejora_todos.csv', index=False)
