==> old_notebooks/8_Test_Estratificado.ipynb <==
import pandas as pd
import numpy as np

def make_mask(total_elems, num_elems):
    random_values = np.random.rand(total_elems)
    
    if total_elems == num_elems:
        return np.full((34), True)
    else:
        border = np.sort(random_values)[num_elems]
        mask = random_values < border
        return mask

# Función para seleccionar aleatoriamente el test estratificado
def get_estratificate_populate(dataframe, num_elems=34):
    finaldf = pd.DataFrame(columns=dataframe.columns)
    
    ratings_values = list(set(dataframe['rating']))
    ratings_values.sort()
    
    for r in ratings_values:
        
        rating_df = dataframe[dataframe['rating'] == r]
        
        # Solo añadimos ratings con representación mayor o igual a num_elems
        if len(rating_df) >= num_elems:
            mask = make_mask(len(rating_df), num_elems)
            finaldf = finaldf.append(rating_df[mask])
    
    return finaldf

all_values = pd.read_csv('data/experiment_data/result_all_values.csv')

final_df = pd.DataFrame(columns=all_values.columns)

for i in range(0, 100):
    estratificado = get_estratificate_populate(all_values)
    final_df = final_df.append(estratificado)
    
final_df

final_df.to_csv('data/experiment_data/result_clean_values.csv', index=False)
