==> old_notebooks/9_Calcular_MAE.ipynb <==
import pandas as pd
import numpy as np

# Cargar Errores

all_errors = pd.read_csv('data/experiment_data/ratings_similar_all_binary.csv')

# Ajustamos Errores

def change_rating(rating):
    if rating == 0.5:
        return 1
    elif rating == 1.5:
        return 2
    elif rating == 2.5:
        return 3
    elif rating == 3.5:
        return 4
    elif rating == 4.5:
        return 5
    else:
        return rating
    
# all_errors['rating'] = all_errors.apply(lambda row: change_rating(row['rating']), axis=1)

def calculate_error(predicted_value, ratings, k):
    
    num_values = min(np.sum(~np.isnan(ratings)), k)
    all_ratings = ratings[:num_values]
    
    result = 0.0
    for r in all_ratings:
        result += abs(predicted_value - r)
        
    return result

def total_values(predicted_value, ratings, k):
    return min(np.sum(~np.isnan(ratings)), k)

k_values = [1, 2, 3, 5, 10]

for k in k_values:
    all_errors['sum_error_k_' + str(k)] = all_errors.apply(lambda row: calculate_error(row['predicted'], row[4:14].values, k), axis=1)
    all_errors['real_k_' + str(k)] = all_errors.apply(lambda row: total_values(row['predicted'], row[4:14].values, k), axis=1)

all_errors

k_values = [1, 2, 3, 5, 10]
ratings = [1.0, 2.0, 3.0, 4.0, 5.0]

results = {}
for k in k_values:
    results_k_1 = {}
    for r in ratings:
        results_k_1['N_rating_' + str(r)] = len(all_errors[all_errors['rating'] == r])

        total_error = np.sum(all_errors[all_errors['rating'] == r]['sum_error_k_' + str(k)].values)
        total_elems = np.sum(all_errors[all_errors['rating'] == r]['real_k_' + str(k)].values)

        results_k_1['mae_rating_' + str(r)] = total_error / total_elems
        
    results['K-' + str(k)] = results_k_1

result_k_1_df = pd.DataFrame(all_errors.groupby(by='rating').count()['userId']).reset_index()
result_k_1_df

def get_error(rating, k):
    sum_error = np.sum(all_errors[all_errors['rating'] == rating]['sum_error_k_{}'.format(k)])
    num_error = np.sum(all_errors[all_errors['rating'] == rating]['real_k_{}'.format(k)])
    return sum_error / num_error

result_k_1_df['error_binary'] = result_k_1_df.apply(lambda row: get_error(row['rating'], 1), axis=1)

result_k_1_df

[5, 37, 4, 85, 34, 295, 95, 304, 66, 219]
